backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0.post1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
